pyproject.toml
[tool.poetry]
name = "markovify_playground"
version = "0.1.0"
description = ""
authors = ["SnO2WMaN <me@sno2wman.net>"]

[tool.poetry.dependencies]
python = "^3.9"
mecab-python3 = "^1.0.5"
unidic-lite = "^1.0.8"
markovify = "^0.9.4"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
